Fix index offset being incorrect if element is coming from outside
https://github.com/flowapp/droppable.js/commit/63a481f2adf10051619538a098327a647c53c472

Fixed a problem with index being too large when `options.manual` was true in
sortable. However, it didn't take into account the possibility that the items
you were sorting were coming from an _outside_ source, so it could be
subtracting when it wasn't actually needed, making the index too _small_.

This checks each element being sorted and only subtracts from the index if the
element is contained in the original container to begin with.