Remove cursorInsideElement check for draggable
This was in place to hack around dragLeave events firing if a droppable element had children. When a user would drag-leave out of a child element, a drag leave would fire on the parent and cause problems.

However, checking the bounds of the window and calculating whether someone dragged out of the droppable area was not entirely accurate either, because if a browser window was in the background and a small window was over top, a user could "leave" into the small on-top window and trick the calculation into thinking that the drag leave didn't happen.

Using pointer-events: none on all nested children also solves the drag leave firing problem, and doesn't the window-on-top bug either.