Check pointer events attribute on drag leave target before handling it
It is possible for a dragLeave event to fire from a child element before the hover class was added on dragEventer and the pointer-events: none style applied.

By the time the event is looked at by droppable js, however, pointer-events: none will had had time to be applied, so we check to make sure the target has pointer-events enabled before cleaning up on dragLeave.