Clear dragEnter state on window mouseout
Since we check client rect offsets in order to decide whether to clear the dragEnter state (since otherwise drag leave events fired by child elements will bubble up and cause false positives) it's possible to actually drag leave into a window on TOP of the browser window and fail to actually clean up the drag and drop state.

This adds a fallback on window dragLeave to make sure we clean up when the user mouseouts from the window.