Update cursorOutsideElement to check if elem is scrolled out of view
The existing check did not take into account elements that were scrolled partially out of view, so even when the cursor was "outside" of the element it was still considered inside if the cursor was positioned over a hidden part of the element outside of the scrollable container.
This updates the method to additionally check that the cursor position, if inside the element, is also within the scrollable container where the element is actually visible.