Update Sortable handleDrop to pass indexAdjusted bool to callback
This is to let the callback know if the placeholder index has been adjusted by -= 1, and so it may be re-adjusted if necessary.

The index would need to be re-adjusted if a view has multiple sortable sections nested within each other and the original element is from the same section as the placeholder - in this case the placeholder index is unaffected by the original element position and therefore does not need to remain adjusted by -= 1.