Update sortable demo to highlight issue with nested links and dragStart
In most cases, the dragStart/context methods were given the correct
elements: those that were targeted via options.itemSelector.  There was
an edge case where if the itemSelector element contains an A element
with an HREF attribute, it would be passed as the dragged element
rather than the parent that matched the selector.

https://app.bananastand.org/organizations/52/teams/1/tasks/5709507